backend/requirements.txt
fastapi
uvicorn
websockets
pandas
scikit-learn
ta
smartapi-python
python-dotenv
pyotp
requests
